Introduce ArchiveStorage trait that group by all logic.
